# More some performance optimizations