Another solid decrease in the diff time (7s => 4s). This plus the last PR have decreased the diff time to about 25% of the original time. I'm now seeing 40s => 10s on one page and 25s => 4s on another overall.